// packages/ui/src/_generated/palette/hues/cyan.stylex.ts
// AUTO-GENERATED by packages/system-palette-codegen — do not edit.
// To regenerate: pnpm codegen:palette

import * as stylex from "@stylexjs/stylex";

export const cyan = stylex.defineConsts({
  _0: "#000000",
  _2: "#000307",
  _5: "#000C14",
  _7: "#00111C",
  _9: "#001623",
  _11: "#001C2A",
  _13: "#002232",
  _20: "#003951",
  _30: "#005C7F",
  _40: "#0083B4",
  _50: "#25A6DE",
  _60: "#46BBF4",
  _70: "#68C9FF",
  _80: "#9BD9FF",
  _90: "#CBE9FF",
  _92: "#D3ECFF",
  _95: "#E0F2FF",
  _97: "#EDF6FF",
  _98: "#F5FAFF",
  _99: "#FBFCFF",
  _100: "#FFFFFF",
});

export const cyan_rgb = stylex.defineConsts({
  _0: "0,0,0",
  _2: "0,3,7",
  _5: "0,12,20",
  _7: "0,17,28",
  _9: "0,22,35",
  _11: "0,28,42",
  _13: "0,34,50",
  _20: "0,57,81",
  _30: "0,92,127",
  _40: "0,131,180",
  _50: "37,166,222",
  _60: "70,187,244",
  _70: "104,201,255",
  _80: "155,217,255",
  _90: "203,233,255",
  _92: "211,236,255",
  _95: "224,242,255",
  _97: "237,246,255",
  _98: "245,250,255",
  _99: "251,252,255",
  _100: "255,255,255",
});