// packages/ui/src/_generated/palette/hues/gray.stylex.ts
// AUTO-GENERATED by packages/system-palette-codegen — do not edit.
// To regenerate: pnpm codegen:palette

import * as stylex from "@stylexjs/stylex";

export const gray = stylex.defineConsts({
  _0: "#000000",
  _2: "#080907",
  _5: "#131412",
  _7: "#1A1B18",
  _9: "#1F201D",
  _11: "#252623",
  _13: "#2B2C29",
  _20: "#424340",
  _30: "#656562",
  _40: "#8C8B88",
  _50: "#ADACA9",
  _60: "#C1C0BD",
  _70: "#D0CECB",
  _80: "#DFDDDA",
  _90: "#EFEDEA",
  _92: "#F2F0ED",
  _95: "#F7F5F2",
  _97: "#FBF9F6",
  _98: "#FEFCF9",
  _99: "#FFFEFB",
  _100: "#FFFFFF",
});

export const gray_rgb = stylex.defineConsts({
  _0: "0,0,0",
  _2: "8,9,7",
  _5: "19,20,18",
  _7: "26,27,24",
  _9: "31,32,29",
  _11: "37,38,35",
  _13: "43,44,41",
  _20: "66,67,64",
  _30: "101,101,98",
  _40: "140,139,136",
  _50: "173,172,169",
  _60: "193,192,189",
  _70: "208,206,203",
  _80: "223,221,218",
  _90: "239,237,234",
  _92: "242,240,237",
  _95: "247,245,242",
  _97: "251,249,246",
  _98: "254,252,249",
  _99: "255,254,251",
  _100: "255,255,255",
});