// packages/ui/src/_generated/palette/hues/green.stylex.ts
// AUTO-GENERATED by packages/system-palette-codegen — do not edit.
// To regenerate: pnpm codegen:palette

import * as stylex from "@stylexjs/stylex";

export const green = stylex.defineConsts({
  _0: "#000000",
  _2: "#000601",
  _5: "#001103",
  _7: "#001804",
  _9: "#001D06",
  _11: "#002408",
  _13: "#002B0B",
  _20: "#004516",
  _30: "#006C27",
  _40: "#00973A",
  _50: "#24BD50",
  _60: "#43D263",
  _70: "#54E170",
  _80: "#65F17D",
  _90: "#90FF9B",
  _92: "#A6FFAB",
  _95: "#C5FFC4",
  _97: "#DBFFD8",
  _98: "#ECFFE8",
  _99: "#F9FFF4",
  _100: "#FFFFFF",
});

export const green_rgb = stylex.defineConsts({
  _0: "0,0,0",
  _2: "0,6,1",
  _5: "0,17,3",
  _7: "0,24,4",
  _9: "0,29,6",
  _11: "0,36,8",
  _13: "0,43,11",
  _20: "0,69,22",
  _30: "0,108,39",
  _40: "0,151,58",
  _50: "36,189,80",
  _60: "67,210,99",
  _70: "84,225,112",
  _80: "101,241,125",
  _90: "144,255,155",
  _92: "166,255,171",
  _95: "197,255,196",
  _97: "219,255,216",
  _98: "236,255,232",
  _99: "249,255,244",
  _100: "255,255,255",
});