// packages/ui/src/_generated/palette/hues/purple.stylex.ts
// AUTO-GENERATED by packages/system-palette-codegen — do not edit.
// To regenerate: pnpm codegen:palette

import * as stylex from "@stylexjs/stylex";

export const purple = stylex.defineConsts({
  _0: "#000000",
  _2: "#040009",
  _5: "#0F0019",
  _7: "#150023",
  _9: "#1C002B",
  _11: "#230035",
  _13: "#2A0040",
  _20: "#480067",
  _30: "#7405A3",
  _40: "#A043D0",
  _50: "#C669F5",
  _60: "#D989FF",
  _70: "#E3A3FF",
  _80: "#EDBEFF",
  _90: "#F7DBFF",
  _92: "#F9E0FF",
  _95: "#FCE9FF",
  _97: "#FFF1FF",
  _98: "#FDF9FF",
  _99: "#FFFCFF",
  _100: "#FFFFFF",
});

export const purple_rgb = stylex.defineConsts({
  _0: "0,0,0",
  _2: "4,0,9",
  _5: "15,0,25",
  _7: "21,0,35",
  _9: "28,0,43",
  _11: "35,0,53",
  _13: "42,0,64",
  _20: "72,0,103",
  _30: "116,5,163",
  _40: "160,67,208",
  _50: "198,105,245",
  _60: "217,137,255",
  _70: "227,163,255",
  _80: "237,190,255",
  _90: "247,219,255",
  _92: "249,224,255",
  _95: "252,233,255",
  _97: "255,241,255",
  _98: "253,249,255",
  _99: "255,252,255",
  _100: "255,255,255",
});